- Data Viewer functioning correctly, - Data Viewer a single instance now instead of one per data source, - Added ability to "purge" state-stored data, - Themes, sub-themes, metadata and data all now integrated
