+++ Frontend/src/actions/types.js
@@ -7,6 +7,7 @@ export const FETCH_CONFIG_REJECTED = 'FETCH_CONFIG_REJECTED';
 export const FETCH_DATA = 'FETCH_DATA';
 export const FETCH_DATA_FULFILLED = 'FETCH_DATA_FULFILLED';
 export const FETCH_DATA_REJECTED = 'FETCH_DATA_REJECTED';
+export const PURGE_DATA = 'PURGE_DATA';
 
 // grid
 export const FETCH_LAYOUT = 'FETCH_LAYOUT';
