+++ Frontend/src/components/data/DataTable/OptionsSidePanel.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { withStyles } from '@material-ui/core/styles';
+
+const styles = ({
+
+});
+
+class OptionsSidePanel extends React.Component {
+  render() {
+    return (
+      <div>OptionsSidePanel</div>
+    )
+  }
+}
+
+export default withStyles(styles)(OptionsSidePanel);
