--- Frontend/src/components/data/DatasourceTable.js
@@ -1,363 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from "prop-types";
-import { withStyles } from "@material-ui/core/styles";
-import classNames from "classnames";
-import Paper from "@material-ui/core/Paper";
-import TablePagination from "@material-ui/core/TablePagination";
-import Table from "@material-ui/core/Table";
-import TableRow from "@material-ui/core/TableRow";
-import TableCell from "@material-ui/core/TableCell";
-import Checkbox from "@material-ui/core/Checkbox";
-import Tooltip from "@material-ui/core/Tooltip";
-import TableSortLabel from "@material-ui/core/TableSortLabel";
-import TableHead from "@material-ui/core/TableHead";
-import TableBody from '@material-ui/core/TableBody';
-import { lighten } from "@material-ui/core/styles/colorManipulator";
-import Typography from "@material-ui/core/Typography";
-import IconButton from "@material-ui/core/IconButton";
-import DeleteIcon from "@material-ui/icons/Delete";
-import FilterListIcon from "@material-ui/icons/FilterList";
-import Toolbar from "@material-ui/core/Toolbar";
-import Collapse from '@material-ui/core/Collapse';
-
-function desc(a, b, orderBy) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
-  }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
-
-function stableSort(array, cmp) {
-  const stabilizedThis = array.map((el, index) => [el, index]);
-  stabilizedThis.sort((a, b) => {
-    const order = cmp(a[0], b[0]);
-    if (order !== 0) return order;
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map(el => el[0]);
-}
-
-function getSorting(order, orderBy) {
-  return order === 'desc' ? (a, b) => desc(a, b, orderBy) : (a, b) => -desc(a, b, orderBy);
-}
-
-const tableStyles = theme => ({
-  root: {
-    width: '100%',
-    marginTop: theme.spacing.unit * 3,
-  },
-  tableWrapper: {
-    overflowX: 'auto',
-  },
-  table: {
-    minWidth: 640,
-  }
-});
-
-class DatasourceTable extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      order: props.meta.order,
-      orderBy: props.meta.orderBy,
-      selected: [],
-      data: [],
-      page: 0,
-      rowsPerPage: 10
-    };
-  }
-
-  componentWillMount() {
-    this.setState({ data: this.props.data });
-  }
-
-  handleSelectAllClick = event => {
-    if (event.target.checked) {
-      this.setState(state => ({ selected: state.data.map(n => n.id) }));
-      return;
-    }
-    this.setState({ selected: [] });
-  };
-
-  handleRequestSort = (event, property) => {
-    const orderBy = property;
-    let order = 'desc';
-
-    if (this.state.orderBy === property && this.state.order === 'desc') {
-      order = 'asc';
-    }
-
-    this.setState({ order, orderBy });
-  };
-
-  handleClick = (event, id) => {
-    const { selected } = this.state;
-    const selectedIndex = selected.indexOf(id);
-    let newSelected = [];
-
-    if (selectedIndex === -1) {
-      newSelected = newSelected.concat(selected, id);
-    } else if (selectedIndex === 0) {
-      newSelected = newSelected.concat(selected.slice(1));
-    } else if (selectedIndex === selected.length - 1) {
-      newSelected = newSelected.concat(selected.slice(0, -1));
-    } else if (selectedIndex > 0) {
-      newSelected = newSelected.concat(
-        selected.slice(0, selectedIndex),
-        selected.slice(selectedIndex + 1),
-      );
-    }
-
-    this.setState({ selected: newSelected });
-  };
-
-  handleChangePage = (event, page) => {
-    this.setState({ page });
-  };
-
-  handleChangeRowsPerPage = event => {
-    this.setState({ rowsPerPage: event.target.value });
-  };
-
-  isSelected = id => this.state.selected.indexOf(id) !== -1;
-
-  generateCells = rowData => {
-    let cells = [];
-
-    this.props.meta.columns.forEach((column, i) => {
-      let cell;
-
-      if (i === 0) {
-        cell = <TableCell
-          key={i + 1}
-          component="th"
-          scope="row"
-          numeric={column.numeric}
-          padding="none"
-        >
-          {rowData[column.id]}
-        </TableCell>
-      } else {
-        cell = <TableCell
-          key={i + 1}
-          numeric={column.numeric}
-          padding="default"
-        >
-          {rowData[column.id]}
-        </TableCell>
-      }
-
-      cells.push(cell);
-    });
-
-    return cells;
-  };
-
-  render() {
-    const { active, classes, meta, name } = this.props;
-    const { order, orderBy, selected, data, page, rowsPerPage } = this.state;
-
-    const emptyRows = rowsPerPage - Math.min(rowsPerPage, data.length - page * rowsPerPage);
-
-    return (
-      <div className={classes.root}>
-        <Collapse in={active} timeout="auto" unmountOnExit>
-          <Paper>
-            <DataSourceTableToolbar numSelected={selected.length} sourceName={name} />
-            <div className={classes.tableWrapper}>
-              <Table className={classes.table} aria-labelledby="tableTitle">
-                <DataSourceTableHead
-                  numSelected={selected.length}
-                  order={order}
-                  orderBy={orderBy}
-                  onSelectAllClick={this.handleSelectAllClick}
-                  onRequestSort={this.handleRequestSort}
-                  columns={meta.columns}
-                  rowCount={data.length}
-                />
-                <TableBody>
-                  {stableSort(data, getSorting(order, orderBy))
-                  .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                  .map((n, i) => {
-                    const key = i + (page * rowsPerPage);
-                    const isSelected = this.isSelected(key);
-
-                    return (
-                      <TableRow
-                        hover
-                        onClick={event => this.handleClick(event, key)}
-                        role="checkbox"
-                        aria-checked={isSelected}
-                        tabIndex={-1}
-                        key={key}
-                        selected={isSelected}
-                      >
-                        <TableCell padding="checkbox">
-                          <Checkbox checked={isSelected} />
-                        </TableCell>
-                        {this.generateCells(n)}
-                      </TableRow>
-                    );
-                  })}
-                  {emptyRows > 0 && (
-                    <TableRow style={{ height: 49 * emptyRows }}>
-                      <TableCell colSpan={6} />
-                    </TableRow>
-                  )}
-                </TableBody>
-              </Table>
-            </div>
-            <TablePagination
-              component="div"
-              count={data.length}
-              rowsPerPage={rowsPerPage}
-              page={page}
-              backIconButtonProps={{ 'aria-label': 'Previous Page' }}
-              nextIconButtonProps={{ 'aria-label': 'Next Page' }}
-              onChangePage={this.handleChangePage}
-              onChangeRowsPerPage={this.handleChangeRowsPerPage}
-            />
-          </Paper>
-        </Collapse>
-      </div>
-    )
-  }
-}
-
-DatasourceTable.propTypes = {
-  classes: PropTypes.object.isRequired,
-};
-
-export default withStyles(tableStyles)(DatasourceTable);
-
-class DataSourceTableHead extends Component {
-  createSortHandler = property => event => {
-    this.props.onRequestSort(event, property);
-  };
-
-  render() {
-    const { onSelectAllClick, order, orderBy, numSelected, columns, rowCount } = this.props;
-
-    return (
-      <TableHead>
-        <TableRow>
-          <TableCell padding="checkbox">
-            <Checkbox
-              indeterminate={numSelected > 0 && numSelected < rowCount}
-              checked={numSelected === rowCount}
-              onChange={onSelectAllClick}
-            />
-          </TableCell>
-          {columns.map((column, i) => {
-            return (
-              <TableCell
-                key={column.id}
-                numeric={column.numeric}
-                padding={(i === 0) ? 'none' : 'default'}
-                sortDirection={orderBy === column.id ? order : false}
-              >
-                <Tooltip
-                  title="Sort"
-                  placement={column.numeric ? 'bottom-end' : 'bottom-start'}
-                  enterDelay={300}
-                >
-                  <TableSortLabel
-                    active={orderBy === column.id}
-                    direction={order}
-                    onClick={this.createSortHandler(column.id)}
-                  >
-                    {column.label}
-                  </TableSortLabel>
-                </Tooltip>
-              </TableCell>
-            );
-          }, this)}
-        </TableRow>
-      </TableHead>
-    )
-  }
-}
-
-DataSourceTableHead.propTypes = {
-  numSelected: PropTypes.number.isRequired,
-  onRequestSort: PropTypes.func.isRequired,
-  onSelectAllClick: PropTypes.func.isRequired,
-  order: PropTypes.string.isRequired,
-  orderBy: PropTypes.string.isRequired,
-  rowCount: PropTypes.number.isRequired,
-};
-
-const toolbarStyles = theme => ({
-  root: {
-    paddingRight: theme.spacing.unit,
-  },
-  highlight:
-  theme.palette.type === 'light'
-  ? {
-    color: theme.palette.secondary.main,
-    backgroundColor: lighten(theme.palette.secondary.light, 0.85),
-  }
-  : {
-    color: theme.palette.text.primary,
-    backgroundColor: theme.palette.secondary.dark,
-  },
-  spacer: {
-    flex: '1 1 100%',
-  },
-  actions: {
-    color: theme.palette.text.secondary,
-  },
-  title: {
-    flex: '0 0 auto',
-    color: 'white',
-  },
-});
-
-class DataSourceTableToolbar extends Component {
-  render() {
-    const { numSelected, classes, sourceName } = this.props;
-
-    return (
-      <Toolbar className={classNames(classes.root, { [classes.highlight]: numSelected > 0, })}>
-        <div className={classes.title}>
-          {numSelected > 0 ? (
-            <Typography color="textPrimary" variant="subtitle1">
-              {numSelected} selected
-            </Typography>
-          ) : (
-            <Typography variant="h6" id="tableTitle">
-              {sourceName}
-            </Typography>
-          )}
-        </div>
-        <div className={classes.spacer} />
-        <div className={classes.actions}>
-          {numSelected > 0 ? (
-            <Tooltip title="Delete">
-              <IconButton aria-label="Delete">
-                <DeleteIcon />
-              </IconButton>
-            </Tooltip>
-          ) : (
-            <Tooltip title="Filter list">
-              <IconButton aria-label="Filter list">
-                <FilterListIcon />
-              </IconButton>
-            </Tooltip>
-          )}
-        </div>
-      </Toolbar>
-    )
-  }
-}
-
-DataSourceTableToolbar.propTypes = {
-  classes: PropTypes.object.isRequired,
-  numSelected: PropTypes.number.isRequired,
-};
-
-DataSourceTableToolbar = withStyles(toolbarStyles)(DataSourceTableToolbar);
